Basic migration to Kotlin 1.3

+++ src/main/kotlin/com/jessecorbett/diskord/HeartbeatManager.kt
@@ -3,7 +3,7 @@ package com.jessecorbett.diskord
 import com.jessecorbett.diskord.api.gateway.GatewayMessage
 
 interface HeartbeatManager {
-    fun start(heartbeatPeriod: Int, sendHeartbeat: () -> Unit, sendAcknowledgement: () -> Unit)
+    fun start(heartbeatPeriod: Long, sendHeartbeat: () -> Unit, sendAcknowledgement: () -> Unit)
 
     fun acceptHeartbeat(message: GatewayMessage)
