+++ src/main/kotlin/com/jessecorbett/diskord/api/gateway/events/Hello.kt
@@ -3,6 +3,6 @@ package com.jessecorbett.diskord.api.gateway.events
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class Hello(
-        @JsonProperty("heartbeat_interval") val heartbeatInterval: Int,
+        @JsonProperty("heartbeat_interval") val heartbeatInterval: Long,
         @JsonProperty("_trace") val trace: List<String>
 )
